+++ src/lib/api/schema/index.ts
@@ -1,6 +1,6 @@
 export * from './atoms';
-export * from './GetProvider.schema';
-export * from './GetProviders.schema';
-export * from './GetProviderAvailability.schema';
-export * from './BookAppointment.schema';
+export * from './_GetProvider.schema';
+export * from './_GetProviders.schema';
+export * from './_GetProviderAvailability.schema';
+export * from './_BookAppointment.schema';
 export * from './shared.schema';
